Ratchet metrics and split notification presentation
