+++ src/services/NotificationManager.ts
@@ -1,10 +1,7 @@
 import { injectable, inject } from 'inversify';
 import {
-  ActionRowBuilder,
-  ButtonStyle,
   Client,
   EmbedBuilder,
-  ButtonBuilder,
   GuildMember,
   Message,
   User,
@@ -25,31 +22,16 @@ import { DetectionResult } from './DetectionOrchestrator';
 import { IDetectionEventsRepository } from '../repositories/DetectionEventsRepository';
 import {
   DetectionEvent,
-  Server,
   VerificationStatus,
   AdminActionType,
   VerificationEvent,
-  DetectionType,
 } from '../repositories/types';
 import { DetectionHistoryFormatter } from '../utils/DetectionHistoryFormatter';
-import type { ReportAIAnalysis, VerificationThreadAnalysisResult } from './GPTService';
+import type { VerificationThreadAnalysisResult } from './GPTService';
 import { getDetectionResponseSettings } from '../utils/detectionResponseSettings';
-import { getVerificationActionFailures } from '../utils/verificationActionFailures';
-import { getCaseResponderSettings } from '../utils/caseResponderSettings';
-import { CASE_STAFF_ROUTING_METADATA_KEY } from './ThreadManager';
-import { isDetectionEventExcludedFromAccounting } from '../utils/detectionEventAccounting';
-import {
-  buildCaseAdminActionsCustomId,
-  buildObservedAdminActionsCustomId,
-} from '../utils/adminActionCustomIds';
+import { NotificationPresentationBuilder } from './NotificationPresentationBuilder';
 
 const VERIFICATION_CHANNEL_NAME = 'verification';
-
-export interface NotificationButton {
-  id: string;
-  label: string;
-  style: ButtonStyle;
-}
 
 /**
  * Interface for NotificationManager service
@@ -136,21 +118,6 @@ export interface INotificationManager {
   ): Promise<boolean>;
 }
 
-interface ThreadAnalysisMetadata {
-  analyzedMessageIds?: unknown;
-  latestAnalysis?: {
-    result: 'likely_legitimate' | 'needs_review' | 'likely_suspicious';
-    confidence: number;
-    summary: string;
-    reasonCodes?: string[];
-    legitimacySignals?: string[];
-    suspicionSignals?: string[];
-    recommendedNextQuestion?: string;
-    recommendedAction?: 'none' | 'ask_followup' | 'manual_review' | 'restrict';
-    analyzedMessageCount: number;
-  };
-}
-
 interface ObservedDetectionMetadata {
   observed_notification_message_id?: string;
   observed_notification_last_notified_at?: string;
@@ -167,9 +134,7 @@ export class NotificationManager implements INotificationManager {
   private client: Client;
   private configService: IConfigService;
   private detectionEventsRepository: IDetectionEventsRepository;
-  private static readonly THREAD_ANALYSIS_FIELD_NAME = 'AI Thread Analysis';
-  private static readonly LATEST_ADMIN_ACTION_FIELD_NAME = 'Latest Admin Action';
-  private static readonly MODERATION_ACTION_WARNING_FIELD_NAME = 'Moderation Action Warning';
+  private presentationBuilder = new NotificationPresentationBuilder();
 
   constructor(
     @inject(TYPES.DiscordClient) client: Client,
@@ -202,15 +167,19 @@ export class NotificationManager implements INotificationManager {
     }
 
     try {
-      // Create the base embed
-      const embed = await this.createSuspiciousUserEmbed(
+      const detectionEvents = await this.detectionEventsRepository.findByServerAndUser(
+        member.guild.id,
+        member.id
+      );
+      const embed = this.presentationBuilder.createSuspiciousUserEmbed(
         member,
         detectionResult,
         verificationEvent,
+        detectionEvents,
         sourceMessage
       );
       const serverConfig = await this.configService.getServerConfig(member.guild.id);
-      const actionRow = this.createActionRow(member.id);
+      const actionRow = this.presentationBuilder.createActionRow(member.id);
 
       // If we have an existing message, update it, otherwise create new
       if (verificationEvent.notification_message_id) {
@@ -225,10 +194,10 @@ export class NotificationManager implements INotificationManager {
       }
 
       // Create a new message
-      const notificationRoleIds = this.getCaseNotificationRoleIds(serverConfig);
+      const notificationRoleIds = this.presentationBuilder.getCaseNotificationRoleIds(serverConfig);
       return await adminChannel.send({
-        content: this.formatRoleMentions(notificationRoleIds),
-        allowedMentions: this.createAdminAllowedMentions(notificationRoleIds),
+        content: this.presentationBuilder.formatRoleMentions(notificationRoleIds),
+        allowedMentions: this.presentationBuilder.createAdminAllowedMentions(notificationRoleIds),
         embeds: [embed],
         components: [actionRow],
       });
@@ -264,7 +233,7 @@ export class NotificationManager implements INotificationManager {
         detectionEvents,
         responseSettings.observedNotificationWindowMinutes
       );
-      const embed = this.createObservedDetectionEmbed(
+      const embed = this.presentationBuilder.createObservedDetectionEmbed(
         member,
         detectionResult,
         detectionEvents,
@@ -272,7 +241,7 @@ export class NotificationManager implements INotificationManager {
       );
       const actionDetectionEventId = detectionResult.detectionEventId ?? detectionEvents[0]?.id;
       const components = actionDetectionEventId
-        ? this.createObservedActionRows(member.id, actionDetectionEventId)
+        ? this.presentationBuilder.createObservedActionRows(member.id, actionDetectionEventId)
         : [];
 
       let notificationMessage: Message | null = null;
@@ -282,7 +251,7 @@ export class NotificationManager implements INotificationManager {
           .catch(() => null);
         if (existingMessage) {
           notificationMessage = await existingMessage.edit({
-            allowedMentions: this.createAdminAllowedMentions(),
+            allowedMentions: this.presentationBuilder.createAdminAllowedMentions(),
             embeds: [embed],
             components,
           });
@@ -290,10 +259,11 @@ export class NotificationManager implements INotificationManager {
       }
 
       if (!notificationMessage) {
-        const notificationRoleIds = this.getCaseNotificationRoleIds(serverConfig);
+        const notificationRoleIds =
+          this.presentationBuilder.getCaseNotificationRoleIds(serverConfig);
         notificationMessage = await notificationChannel.send({
-          content: this.formatRoleMentions(notificationRoleIds),
-          allowedMentions: this.createAdminAllowedMentions(notificationRoleIds),
+          content: this.presentationBuilder.formatRoleMentions(notificationRoleIds),
+          allowedMentions: this.presentationBuilder.createAdminAllowedMentions(notificationRoleIds),
           embeds: [embed],
           components,
         });
@@ -358,19 +328,17 @@ export class NotificationManager implements INotificationManager {
 
       const updatedEmbed = EmbedBuilder.from(message.embeds[0]);
       const timestamp = Math.floor(Date.now() / 1000);
-      const field = {
-        name: 'Action Taken',
-        value: `<@${admin.id}> ${actionDescription} <t:${timestamp}:R>`,
-        inline: false,
-      };
-      const existingFields =
-        updatedEmbed.data.fields?.filter(
-          (existingField) =>
-            existingField.name !== field.name && existingField.name !== 'Action Reverted'
-        ) ?? [];
-      updatedEmbed.setFields(...existingFields, field);
+      this.presentationBuilder.addObservedActionTakenField(
+        updatedEmbed,
+        actionDescription,
+        admin.id,
+        timestamp
+      );
 
-      const components = this.createObservedActionRows(detectionEvent.user_id, detectionEvent.id);
+      const components = this.presentationBuilder.createObservedActionRows(
+        detectionEvent.user_id,
+        detectionEvent.id
+      );
 
       await message.edit({
         allowedMentions: { parse: [] },
@@ -417,7 +385,10 @@ export class NotificationManager implements INotificationManager {
         return false;
       }
 
-      const components = this.createObservedActionRows(detectionEvent.user_id, detectionEvent.id);
+      const components = this.presentationBuilder.createObservedActionRows(
+        detectionEvent.user_id,
+        detectionEvent.id
+      );
       if (!message.embeds.length) {
         await message.edit({ allowedMentions: { parse: [] }, components });
         return true;
@@ -425,17 +396,12 @@ export class NotificationManager implements INotificationManager {
 
       const updatedEmbed = EmbedBuilder.from(message.embeds[0]);
       const timestamp = Math.floor(Date.now() / 1000);
-      const field = {
-        name: 'Action Reverted',
-        value: `<@${admin.id}> ${actionDescription} <t:${timestamp}:R>`,
-        inline: false,
-      };
-      const existingFields =
-        updatedEmbed.data.fields?.filter(
-          (existingField) =>
-            existingField.name !== 'Action Taken' && existingField.name !== field.name
-        ) ?? [];
-      updatedEmbed.setFields(...existingFields, field);
+      this.presentationBuilder.addObservedActionRevertedField(
+        updatedEmbed,
+        actionDescription,
+        admin.id,
+        timestamp
+      );
 
       await message.edit({
         allowedMentions: { parse: [] },
@@ -475,56 +441,14 @@ export class NotificationManager implements INotificationManager {
       // Create a new embed based on the existing one
       const updatedEmbed = EmbedBuilder.from(existingEmbed);
 
-      const timestamp = Math.floor(Date.now() / 1000);
-      const actionLogField = updatedEmbed.data.fields?.find((field) => field.name === 'Action Log');
-      this.upsertLatestAdminActionField(updatedEmbed, actionTaken, admin.id, timestamp);
-
-      let actionLogContent = `• ${this.formatAdminActionEvent(actionTaken, admin.id, timestamp)}`;
-
-      // If a thread was created, add/update a dedicated thread link field
-      if (thread && actionTaken === AdminActionType.CREATE_THREAD) {
-        const threadField = {
-          name: 'Verification Thread',
-          value: `[Click here to view the thread](${thread.url})`,
-          inline: false,
-        };
-        // Check if thread field already exists
-        const existingThreadFieldIndex = updatedEmbed.data.fields?.findIndex(
-          (field) => field.name === 'Verification Thread'
-        );
-        if (existingThreadFieldIndex !== undefined && existingThreadFieldIndex > -1) {
-          // Update existing field
-          updatedEmbed.spliceFields(existingThreadFieldIndex, 1, threadField);
-        } else {
-          // Add new field
-          updatedEmbed.addFields(threadField);
-        }
-      }
-
-      // Update the thread status if it's a verification or ban action
-      if (message.guildId && actionTaken === AdminActionType.VERIFY) {
-        // Update embed color based on resolution
-        updatedEmbed.setColor(0x00ff00);
-      }
-
-      // Update the thread status if it's a verification or ban action
-      if (message.guildId && actionTaken === AdminActionType.BAN) {
-        // Update embed color based on resolution
-        updatedEmbed.setColor(0x000000);
-      }
-
-      if (actionLogField) {
-        // Append to existing log
-        actionLogContent = `${actionLogField.value}\n${actionLogContent}`;
-      }
-
-      if (actionLogField) {
-        // Update existing field
-        actionLogField.value = actionLogContent;
-      } else {
-        // Add new field
-        updatedEmbed.addFields({ name: 'Action Log', value: actionLogContent, inline: false });
-      }
+      this.presentationBuilder.upsertAdminActionLog(
+        updatedEmbed,
+        actionTaken,
+        admin.id,
+        Math.floor(Date.now() / 1000),
+        thread?.url,
+        Boolean(message.guildId)
+      );
 
       // Update the message embed. Let button updates be handled separately.
       await message.edit({ allowedMentions: { parse: [] }, embeds: [updatedEmbed] });
@@ -545,26 +469,11 @@ export class NotificationManager implements INotificationManager {
       const existingEmbed = message.embeds[0];
       const updatedEmbed = EmbedBuilder.from(existingEmbed);
 
-      const value = this.formatThreadAnalysisFieldValue({
-        ...analysis,
-        analyzedMessageCount,
-      });
-
-      const field = {
-        name: NotificationManager.THREAD_ANALYSIS_FIELD_NAME,
-        value,
-        inline: false,
-      };
-
-      const existingFieldIndex = updatedEmbed.data.fields?.findIndex(
-        (embedField) => embedField.name === field.name
+      this.presentationBuilder.upsertThreadAnalysisField(
+        updatedEmbed,
+        analysis,
+        analyzedMessageCount
       );
-
-      if (existingFieldIndex !== undefined && existingFieldIndex > -1) {
-        updatedEmbed.spliceFields(existingFieldIndex, 1, field);
-      } else {
-        updatedEmbed.addFields(field);
-      }
 
       await message.edit({ allowedMentions: { parse: [] }, embeds: [updatedEmbed] });
       return true;
@@ -574,146 +483,12 @@ export class NotificationManager implements INotificationManager {
     }
   }
 
-  private truncateEmbedFieldValue(value: string): string {
-    const maxLength = 1024;
-    if (value.length <= maxLength) {
-      return value;
-    }
-
-    return `${value.slice(0, maxLength - 3)}...`;
-  }
-
-  private formatAdminActionLabel(actionTaken: AdminActionType): string {
-    switch (actionTaken) {
-      case AdminActionType.BAN:
-        return 'Banned';
-      case AdminActionType.VERIFY:
-        return 'Verified';
-      case AdminActionType.CREATE_THREAD:
-        return 'Created verification thread';
-      case AdminActionType.REOPEN:
-        return 'Reopened verification';
-      case AdminActionType.RESTRICT:
-        return 'Restricted';
-      case AdminActionType.OPEN_CASE:
-        return 'Opened case';
-      case AdminActionType.DISMISS:
-        return 'Dismissed alert';
-      case AdminActionType.FALSE_POSITIVE:
-        return 'Marked false positive';
-      case AdminActionType.UNDO_OBSERVED_ACTION:
-        return 'Undid observed action';
-      case AdminActionType.REJECT:
-        return 'Rejected';
-      default:
-        return actionTaken;
-    }
-  }
-
-  private formatAdminActionEvent(
-    actionTaken: AdminActionType,
-    adminId: string,
-    timestamp: number
-  ): string {
-    return `${this.formatAdminActionLabel(actionTaken)} by <@${adminId}> at <t:${timestamp}:F>`;
-  }
-
-  private upsertLatestAdminActionField(
-    embed: EmbedBuilder,
-    actionTaken: AdminActionType,
-    adminId: string,
-    timestamp: number
-  ): void {
-    const field = {
-      name: NotificationManager.LATEST_ADMIN_ACTION_FIELD_NAME,
-      value: this.formatAdminActionEvent(actionTaken, adminId, timestamp),
-      inline: false,
-    };
-    const fields = (embed.data.fields ?? []).filter(
-      (existingField) => existingField.name !== field.name
-    );
-    const confidenceIndex = fields.findIndex(
-      (existingField) => existingField.name === 'Detection Confidence'
-    );
-    const insertIndex = confidenceIndex >= 0 ? confidenceIndex + 1 : 0;
-    fields.splice(insertIndex, 0, field);
-    embed.setFields(...fields);
-  }
-
-  private upsertLatestResolutionField(
-    embed: EmbedBuilder,
-    verificationEvent: VerificationEvent
-  ): void {
-    if (!verificationEvent.resolved_by || !verificationEvent.resolved_at) {
-      return;
-    }
-
-    const actionTaken =
-      verificationEvent.status === VerificationStatus.BANNED
-        ? AdminActionType.BAN
-        : verificationEvent.status === VerificationStatus.VERIFIED
-          ? AdminActionType.VERIFY
-          : null;
-    if (!actionTaken) {
-      return;
-    }
-
-    this.upsertLatestAdminActionField(
-      embed,
-      actionTaken,
-      verificationEvent.resolved_by,
-      Math.floor(verificationEvent.resolved_at.getTime() / 1000)
-    );
-  }
-
   private metadataToRecord(metadata: DetectionEvent['metadata']): Record<string, unknown> {
     if (!metadata || typeof metadata !== 'object' || Array.isArray(metadata)) {
       return {};
     }
 
     return { ...metadata } as Record<string, unknown>;
-  }
-
-  private createAdminAllowedMentions(roleIds?: string[] | string | null): {
-    parse: [];
-    roles: string[];
-    users: [];
-    repliedUser: false;
-  } {
-    const roles = Array.isArray(roleIds)
-      ? roleIds
-      : typeof roleIds === 'string' && roleIds
-        ? [roleIds]
-        : [];
-
-    return {
-      parse: [],
-      roles,
-      users: [],
-      repliedUser: false,
-    };
-  }
-
-  private getCaseNotificationRoleIds(serverConfig: Server): string[] {
-    const roleIds = new Set<string>();
-    if (serverConfig.admin_notification_role_id) {
-      roleIds.add(serverConfig.admin_notification_role_id);
-    }
-
-    const responderSettings = getCaseResponderSettings(serverConfig.settings);
-    if (responderSettings.routingMode !== 'off') {
-      responderSettings.roleIds.forEach((roleId) => roleIds.add(roleId));
-    }
-
-    return [...roleIds];
-  }
-
-  private formatRoleMentions(roleIds: readonly string[]): string | undefined {
-    if (roleIds.length === 0) {
-      return undefined;
-    }
-
-    return roleIds.map((roleId) => `<@&${roleId}>`).join(' ');
   }
 
   private findRecentObservedDetectionNotification(
@@ -765,122 +540,6 @@ export class NotificationManager implements INotificationManager {
     return (await this.configService.getAdminChannel(guildId)) ?? null;
   }
 
-  private formatDetectionTrigger(
-    detectionResult: DetectionResult,
-    sourceMessage?: Message
-  ): string {
-    if (detectionResult.triggerSource === DetectionType.SUSPICIOUS_CONTENT) {
-      const safeContent = detectionResult.triggerContent
-        ? `\`${detectionResult.triggerContent}\``
-        : '`Message content unavailable`';
-      return sourceMessage
-        ? `[Observed message](${sourceMessage.url}): ${safeContent}`
-        : `Observed message: ${safeContent}`;
-    }
-
-    if (detectionResult.triggerSource === DetectionType.NEW_ACCOUNT) {
-      return 'Observed upon joining server';
-    }
-
-    if (detectionResult.triggerSource === DetectionType.USER_REPORT) {
-      return `Observed via user report: \`${detectionResult.triggerContent || 'No reason provided'}\``;
-    }
-
-    if (detectionResult.triggerSource === DetectionType.GPT_ANALYSIS) {
-      return `Observed via manual review: \`${detectionResult.triggerContent || 'Manual flag'}\``;
-    }
-
-    return 'Observed suspicious activity';
-  }
-
-  private createObservedDetectionEmbed(
-    member: GuildMember,
-    detectionResult: DetectionResult,
-    detectionEvents: DetectionEvent[],
-    sourceMessage?: Message
-  ): EmbedBuilder {
-    const accountCreatedAt = new Date(member.user.createdTimestamp);
-    const accountCreatedTimestamp = Math.floor(accountCreatedAt.getTime() / 1000);
-    const joinedServerTimestamp = member.joinedAt
-      ? Math.floor(member.joinedAt.getTime() / 1000)
-      : null;
-    const confidencePercent = Math.round(detectionResult.confidence * 100);
-    const reasonsFormatted = detectionResult.reasons.map((reason) => `• ${reason}`).join('\n');
-    const recentEvents = detectionEvents.slice(0, 5);
-    const detectionHistory = recentEvents
-      .map((event) => {
-        const timestamp = Math.floor(new Date(event.detected_at).getTime() / 1000);
-        return `• <t:${timestamp}:R>: ${event.detection_type} (${Math.round(event.confidence * 100)}% confidence)${this.formatAccountingSuffix(event)}`;
-      })
-      .join('\n');
-
-    const embed = new EmbedBuilder()
-      .setColor(0xffc107)
-      .setTitle('Suspicious Activity Observed')
-      .setDescription(
-        `Drasil observed suspicious activity from <@${member.id}>. No automatic restriction was applied.`
-      )
-      .setThumbnail(member.user.displayAvatarURL())
-      .addFields(
-        { name: 'Username', value: member.user.tag, inline: true },
-        { name: 'User ID', value: member.id, inline: true },
-        {
-          name: 'Account Created',
-          value: `<t:${accountCreatedTimestamp}:F> (<t:${accountCreatedTimestamp}:R>)`,
-          inline: false,
-        },
-        {
-          name: 'Joined Server',
-          value: joinedServerTimestamp
-            ? `<t:${joinedServerTimestamp}:F> (<t:${joinedServerTimestamp}:R>)`
-            : 'Unknown',
-          inline: false,
-        },
-        { name: 'Detection Confidence', value: `${confidencePercent}%`, inline: true },
-        {
-          name: 'Trigger',
-          value: this.truncateEmbedFieldValue(
-            this.formatDetectionTrigger(detectionResult, sourceMessage)
-          ),
-        },
-        {
-          name: 'Reasons',
-          value: this.truncateEmbedFieldValue(reasonsFormatted || 'No specific reason provided'),
-        }
-      )
-      .setTimestamp();
-
-    const aiDiagnosticFieldValue = this.formatGptDiagnosticFieldValue(detectionResult);
-    if (aiDiagnosticFieldValue) {
-      embed.addFields({
-        name: 'AI Analysis',
-        value: aiDiagnosticFieldValue,
-        inline: false,
-      });
-    }
-
-    const reportAiFieldValue = this.formatReportAiFieldValue(
-      detectionResult.reportAiAnalysis ??
-        this.findReportAiAnalysis(detectionEvents, detectionResult.detectionEventId)
-    );
-    if (reportAiFieldValue) {
-      embed.addFields({
-        name: 'AI Report Triage',
-        value: reportAiFieldValue,
-        inline: false,
-      });
-    }
-
-    if (detectionHistory) {
-      embed.addFields({
-        name: 'Recent Detection History',
-        value: this.truncateEmbedFieldValue(detectionHistory),
-      });
-    }
-
-    return embed;
-  }
-
   private async getMessageForVerificationEvent(
     verificationEvent: VerificationEvent
   ): Promise<Message> {
@@ -895,516 +554,6 @@ export class NotificationManager implements INotificationManager {
     }
 
     return await adminChannel.messages.fetch(verificationEvent.notification_message_id);
-  }
-
-  /**
-   * Creates an embed for displaying suspicious user information
-   * @param member The guild member
-   * @param detectionResult The detection result
-   * @param sourceMessage Optional message that triggered the detection
-   * @returns An EmbedBuilder with user information
-   */
-  private async createSuspiciousUserEmbed(
-    member: GuildMember,
-    detectionResult: DetectionResult,
-    verificationEvent: VerificationEvent,
-    sourceMessage?: Message
-  ): Promise<EmbedBuilder> {
-    const accountCreatedAt = new Date(member.user.createdTimestamp);
-    const joinedServerAt = member.joinedAt;
-
-    // Get unix timestamps for Discord timestamp formatting
-    const accountCreatedTimestamp = Math.floor(accountCreatedAt.getTime() / 1000);
-    const joinedServerTimestamp = joinedServerAt ? Math.floor(joinedServerAt.getTime() / 1000) : 0;
-
-    // Format the account timestamps with both absolute and relative format
-    const accountCreatedFormatted = `<t:${accountCreatedTimestamp}:F> (<t:${accountCreatedTimestamp}:R>)`;
-    const joinedServerFormatted = joinedServerAt
-      ? `<t:${joinedServerTimestamp}:F> (<t:${joinedServerTimestamp}:R>)`
-      : 'Unknown';
-
-    // Convert confidence to Low/Medium/High
-    const confidencePercent = detectionResult.confidence * 100;
-    let confidenceLevel: string;
-    let embedColor: number = 0xff0000; // Default red for suspicious/unverified users
-
-    if (confidencePercent <= 40) {
-      confidenceLevel = '🟢 Low';
-    } else if (confidencePercent <= 70) {
-      confidenceLevel = '🟡 Medium';
-    } else {
-      confidenceLevel = '🔴 High';
-    }
-
-    // Format reasons as bullet points
-    const reasonsFormatted = detectionResult.reasons.map((reason) => `• ${reason}`).join('\n');
-
-    // Create trigger information
-    let triggerInfo: string;
-    if (detectionResult.triggerSource === DetectionType.SUSPICIOUS_CONTENT) {
-      const safeContent = detectionResult.triggerContent
-        ? `\`${detectionResult.triggerContent}\``
-        : '`Message content unavailable`';
-
-      if (sourceMessage) {
-        triggerInfo = `[Flagged for message](${sourceMessage.url}): ${safeContent}`;
-      } else {
-        triggerInfo = `Flagged for message: ${safeContent}`;
-      }
-    } else if (detectionResult.triggerSource === DetectionType.USER_REPORT) {
-      const safeContent = detectionResult.triggerContent
-        ? `\`${detectionResult.triggerContent}\``
-        : '`No report reason provided`';
-      triggerInfo = `Flagged via user report: ${safeContent}`;
-    } else if (detectionResult.triggerSource === DetectionType.GPT_ANALYSIS) {
-      const safeContent = detectionResult.triggerContent
-        ? `\`${detectionResult.triggerContent}\``
-        : '`Manual flag`';
-      triggerInfo = `Flagged via manual review: ${safeContent}`;
-    } else if (detectionResult.triggerSource === DetectionType.NEW_ACCOUNT) {
-      triggerInfo = 'Flagged upon joining server';
-    } else {
-      triggerInfo = 'Flagged for suspicious activity';
-    }
-
-    // Get all detection events for this user in this server
-    const detectionEvents = await this.detectionEventsRepository.findByServerAndUser(
-      member.guild.id,
-      member.id
-    );
-    const countedDetectionEvents = detectionEvents.filter(
-      (event) => !isDetectionEventExcludedFromAccounting(event)
-    );
-
-    // Format detection history
-    let detectionHistory = '';
-    if (detectionEvents.length > 0) {
-      // Sort events by date, most recent first
-      const sortedEvents = detectionEvents.sort(
-        (a, b) => new Date(b.detected_at).getTime() - new Date(a.detected_at).getTime()
-      );
-
-      // Take the 5 most recent events
-      const recentEvents = sortedEvents.slice(0, 5);
-
-      // Format the recent events
-      detectionHistory = recentEvents
-        .map((event) => {
-          const timestamp = Math.floor(new Date(event.detected_at).getTime() / 1000);
-          let entry = `• <t:${timestamp}:R>: ${event.detection_type}`;
-          if (event.message_id) {
-            entry += ` - [View Message](https://discord.com/channels/${member.guild.id}/${event.channel_id}/${event.message_id})`;
-          }
-          entry += ` (${(event.confidence * 100).toFixed(0)}% confidence)`;
-          entry += this.formatAccountingSuffix(event);
-          return entry;
-        })
-        .join('\n');
-
-      // If there are more events, add a count
-      if (sortedEvents.length > 5) {
-        detectionHistory += `\n\n*${sortedEvents.length - 5} more events not shown*`;
-      }
-    }
-
-    // Update embed color based on verification status if thread exists
-    if (verificationEvent.status === VerificationStatus.VERIFIED) {
-      embedColor = 0x00ff00; // Green for verified users
-    } else if (verificationEvent.status === VerificationStatus.BANNED) {
-      embedColor = 0x000000; // Black for banned users
-    }
-
-    // Create the embed
-    const embed = new EmbedBuilder()
-      .setColor(embedColor)
-      .setTitle('Suspicious User Detected')
-      .setDescription(
-        countedDetectionEvents.length > 1
-          ? `<@${member.id}> has been flagged as suspicious ${countedDetectionEvents.length} times.`
-          : `<@${member.id}> has been flagged as suspicious.`
-      )
-      .setThumbnail(member.user.displayAvatarURL())
-      .addFields(
-        { name: 'Username', value: member.user.tag, inline: true },
-        { name: 'User ID', value: member.id, inline: true },
-        { name: 'Account Created', value: accountCreatedFormatted, inline: false },
-        { name: 'Joined Server', value: joinedServerFormatted, inline: false },
-        { name: 'Detection Confidence', value: confidenceLevel, inline: true },
-        { name: 'Trigger', value: triggerInfo, inline: false },
-        { name: 'Reasons', value: reasonsFormatted || 'No specific reason provided', inline: false }
-      )
-      .setTimestamp();
-
-    this.upsertLatestResolutionField(embed, verificationEvent);
-
-    const aiDiagnosticFieldValue = this.formatGptDiagnosticFieldValue(detectionResult);
-    if (aiDiagnosticFieldValue) {
-      embed.addFields({
-        name: 'AI Analysis',
-        value: aiDiagnosticFieldValue,
-        inline: false,
-      });
-    }
-
-    const reportAiFieldValue = this.formatReportAiFieldValue(
-      detectionResult.reportAiAnalysis ??
-        this.findReportAiAnalysis(detectionEvents, detectionResult.detectionEventId)
-    );
-    if (reportAiFieldValue) {
-      embed.addFields({
-        name: 'AI Report Triage',
-        value: reportAiFieldValue,
-        inline: false,
-      });
-    }
-
-    const actionFailureFieldValue = this.formatVerificationActionFailureFieldValue(
-      verificationEvent.metadata
-    );
-    if (actionFailureFieldValue) {
-      embed.addFields({
-        name: NotificationManager.MODERATION_ACTION_WARNING_FIELD_NAME,
-        value: actionFailureFieldValue,
-        inline: false,
-      });
-    }
-
-    const caseStaffWarningFieldValue = this.formatCaseStaffRoutingWarningFieldValue(
-      verificationEvent.metadata
-    );
-    if (caseStaffWarningFieldValue) {
-      embed.addFields({
-        name: 'Case Staff Routing Warning',
-        value: caseStaffWarningFieldValue,
-        inline: false,
-      });
-    }
-
-    // Add detection history if we have any
-    if (detectionHistory) {
-      embed.addFields({
-        name: 'Detection History',
-        value: detectionHistory,
-        inline: false,
-      });
-    }
-
-    // Add verification thread status if it exists
-    if (verificationEvent.thread_id) {
-      const threadStatus =
-        verificationEvent.status === VerificationStatus.VERIFIED ||
-        verificationEvent.status === VerificationStatus.BANNED
-          ? `${verificationEvent.status} by <@${verificationEvent.resolved_by}>`
-          : 'pending'; // Use 'pending' for unresolved states
-      embed.addFields({
-        name: 'Verification Status',
-        value: `[Thread](https://discord.com/channels/${member.guild.id}/${verificationEvent.thread_id}) status: ${threadStatus}`,
-        inline: false,
-      });
-    }
-
-    const persistedThreadAnalysis = this.getThreadAnalysisMetadata(verificationEvent.metadata);
-    if (persistedThreadAnalysis?.latestAnalysis) {
-      embed.addFields({
-        name: NotificationManager.THREAD_ANALYSIS_FIELD_NAME,
-        value: this.formatThreadAnalysisFieldValue(persistedThreadAnalysis.latestAnalysis),
-        inline: false,
-      });
-    }
-
-    return embed;
-  }
-
-  private formatVerificationActionFailureFieldValue(metadata: unknown): string | null {
-    const failures = getVerificationActionFailures(metadata);
-    if (failures.length === 0) {
-      return null;
-    }
-
-    const value = failures
-      .slice(-3)
-      .map((failure) => {
-        const timestamp = Math.floor(new Date(failure.at).getTime() / 1000);
-        const action =
-          failure.action === 'restrict'
-            ? 'Apply restricted role'
-            : failure.action === 'private_evidence_thread'
-              ? 'Create private evidence thread'
-              : 'Create case thread';
-        const when = Number.isFinite(timestamp) ? ` <t:${timestamp}:R>` : '';
-        return `Warning: ${action} failed${when}: ${failure.message}`;
-      })
-      .join('\n');
-
-    return this.truncateEmbedFieldValue(
-      `${value}\nCase record was still created so moderators can review and fix permissions.`
-    );
-  }
-
-  private upsertVerificationActionFailureField(
-    embed: EmbedBuilder,
-    verificationEvent: VerificationEvent
-  ): void {
-    const actionFailureFieldValue = this.formatVerificationActionFailureFieldValue(
-      verificationEvent.metadata
-    );
-    const fieldIndex = embed.data.fields?.findIndex(
-      (field) => field.name === NotificationManager.MODERATION_ACTION_WARNING_FIELD_NAME
-    );
-
-    if (actionFailureFieldValue) {
-      const field = {
-        name: NotificationManager.MODERATION_ACTION_WARNING_FIELD_NAME,
-        value: actionFailureFieldValue,
-        inline: false,
-      };
-      if (fieldIndex !== undefined && fieldIndex >= 0) {
-        embed.spliceFields(fieldIndex, 1, field);
-      } else {
-        embed.addFields(field);
-      }
-      return;
-    }
-
-    if (fieldIndex !== undefined && fieldIndex >= 0) {
-      embed.spliceFields(fieldIndex, 1);
-    }
-  }
-
-  private formatCaseStaffRoutingWarningFieldValue(metadata: unknown): string | null {
-    if (!metadata || typeof metadata !== 'object' || Array.isArray(metadata)) {
-      return null;
-    }
-
-    const routing = (metadata as Record<string, unknown>)[CASE_STAFF_ROUTING_METADATA_KEY];
-    if (!routing || typeof routing !== 'object' || Array.isArray(routing)) {
-      return null;
-    }
-
-    const warnings = (routing as { warnings?: unknown }).warnings;
-    if (!Array.isArray(warnings) || warnings.length === 0) {
-      return null;
-    }
-
-    return this.truncateEmbedFieldValue(
-      warnings
-        .filter((warning): warning is string => typeof warning === 'string' && warning.length > 0)
-        .slice(-3)
-        .map((warning) => `Warning: ${warning}`)
-        .join('\n')
-    );
-  }
-
-  private formatThreadAnalysisFieldValue(analysis: {
-    result: 'likely_legitimate' | 'needs_review' | 'likely_suspicious';
-    confidence: number;
-    summary: string;
-    reasonCodes?: string[];
-    legitimacySignals?: string[];
-    suspicionSignals?: string[];
-    recommendedNextQuestion?: string;
-    recommendedAction?: 'none' | 'ask_followup' | 'manual_review' | 'restrict';
-    analyzedMessageCount: number;
-  }): string {
-    const confidencePercent = Math.round(analysis.confidence * 100);
-    const lines = [
-      `Result: **${analysis.result}** (${confidencePercent}% confidence)`,
-      `Analyzed responses: ${analysis.analyzedMessageCount}`,
-      `Summary: ${analysis.summary}`,
-    ];
-    if (analysis.reasonCodes?.length) {
-      lines.push(`Reason codes: ${analysis.reasonCodes.join(', ')}`);
-    }
-    if (analysis.legitimacySignals?.length) {
-      lines.push(`Legitimacy signals: ${analysis.legitimacySignals.join('; ')}`);
-    }
-    if (analysis.suspicionSignals?.length) {
-      lines.push(`Suspicion signals: ${analysis.suspicionSignals.join('; ')}`);
-    }
-    if (analysis.recommendedNextQuestion) {
-      lines.push(`Next question: ${analysis.recommendedNextQuestion}`);
-    }
-    if (analysis.recommendedAction) {
-      lines.push(`Recommended action: ${analysis.recommendedAction}`);
-    }
-
-    return this.truncateEmbedFieldValue(lines.join('\n'));
-  }
-
-  private findReportAiAnalysis(
-    detectionEvents: DetectionEvent[],
-    detectionEventId?: string
-  ): ReportAIAnalysis | null {
-    const candidates = detectionEventId
-      ? detectionEvents.filter((event) => event.id === detectionEventId)
-      : detectionEvents;
-
-    for (const event of candidates) {
-      const metadata = this.metadataToRecord(event.metadata);
-      const reportAi = metadata.report_ai;
-      if (reportAi && typeof reportAi === 'object' && !Array.isArray(reportAi)) {
-        return reportAi as ReportAIAnalysis;
-      }
-    }
-
-    return null;
-  }
-
-  private formatReportAiFieldValue(analysis: ReportAIAnalysis | null): string | null {
-    if (!analysis) {
-      return null;
-    }
-
-    const confidencePercent = Math.round(analysis.confidence * 100);
-    const lines = [
-      `Result: **${analysis.result}** (${confidencePercent}% confidence)`,
-      `Summary: ${analysis.summary}`,
-      `Recommended action: ${analysis.recommendedAction}`,
-      `Images analyzed: ${analysis.analyzedImageCount}`,
-    ];
-    if (analysis.reasonCodes.length) {
-      lines.push(`Reason codes: ${analysis.reasonCodes.join(', ')}`);
-    }
-    if (analysis.evidenceCategories.length) {
-      lines.push(`Evidence: ${analysis.evidenceCategories.join(', ')}`);
-    }
-    if (analysis.concerns.length) {
-      lines.push(`Concerns: ${analysis.concerns.join('; ')}`);
-    }
-
-    return this.truncateEmbedFieldValue(lines.join('\n'));
-  }
-
-  private formatGptDiagnosticFieldValue(detectionResult: DetectionResult): string | null {
-    const analysis = detectionResult.gptAnalysis;
-    if (!analysis) {
-      return null;
-    }
-
-    const confidencePercent = Math.round(analysis.confidence * 100);
-    const reasonCodes = analysis.reasonCodes.length ? analysis.reasonCodes.join(', ') : 'none';
-    const resultLine = analysis.isFallback
-      ? 'Result: **Unavailable**'
-      : `Result: **${analysis.result}** (${confidencePercent}% confidence)`;
-    return this.truncateEmbedFieldValue(
-      [
-        resultLine,
-        `Primary signal: ${analysis.primarySignal}`,
-        `Reason codes: ${reasonCodes}`,
-        `Summary: ${analysis.summary}`,
-      ].join('\n')
-    );
-  }
-
-  private formatAccountingSuffix(event: DetectionEvent): string {
-    return isDetectionEventExcludedFromAccounting(event) ? ' - ignored for future accounting' : '';
-  }
-
-  private getThreadAnalysisMetadata(
-    metadata: VerificationEvent['metadata']
-  ): ThreadAnalysisMetadata | null {
-    if (!metadata || typeof metadata !== 'object' || Array.isArray(metadata)) {
-      return null;
-    }
-
-    const threadAnalysis = (metadata as { thread_analysis?: unknown }).thread_analysis;
-    if (!threadAnalysis || typeof threadAnalysis !== 'object' || Array.isArray(threadAnalysis)) {
-      return null;
-    }
-
-    const metadataRecord = threadAnalysis as ThreadAnalysisMetadata;
-    const latestAnalysis =
-      metadataRecord.latestAnalysis &&
-      typeof metadataRecord.latestAnalysis === 'object' &&
-      !Array.isArray(metadataRecord.latestAnalysis)
-        ? (metadataRecord.latestAnalysis as Record<string, unknown>)
-        : null;
-    if (!latestAnalysis) {
-      return metadataRecord;
-    }
-
-    const rawResult = latestAnalysis.result;
-    const result =
-      rawResult === 'likely_legitimate' ||
-      rawResult === 'needs_review' ||
-      rawResult === 'likely_suspicious'
-        ? rawResult
-        : rawResult === 'OK'
-          ? 'likely_legitimate'
-          : rawResult === 'SUSPICIOUS'
-            ? 'likely_suspicious'
-            : null;
-
-    if (
-      !result ||
-      typeof latestAnalysis.confidence !== 'number' ||
-      typeof latestAnalysis.summary !== 'string' ||
-      typeof latestAnalysis.analyzedMessageCount !== 'number'
-    ) {
-      return metadataRecord;
-    }
-
-    return {
-      ...metadataRecord,
-      latestAnalysis: {
-        result,
-        confidence: latestAnalysis.confidence,
-        summary: latestAnalysis.summary,
-        reasonCodes: Array.isArray(latestAnalysis.reasonCodes)
-          ? latestAnalysis.reasonCodes.filter((value): value is string => typeof value === 'string')
-          : [],
-        legitimacySignals: Array.isArray(latestAnalysis.legitimacySignals)
-          ? latestAnalysis.legitimacySignals.filter(
-              (value): value is string => typeof value === 'string'
-            )
-          : [],
-        suspicionSignals: Array.isArray(latestAnalysis.suspicionSignals)
-          ? latestAnalysis.suspicionSignals.filter(
-              (value): value is string => typeof value === 'string'
-            )
-          : [],
-        recommendedNextQuestion:
-          typeof latestAnalysis.recommendedNextQuestion === 'string'
-            ? latestAnalysis.recommendedNextQuestion
-            : undefined,
-        recommendedAction:
-          latestAnalysis.recommendedAction === 'none' ||
-          latestAnalysis.recommendedAction === 'ask_followup' ||
-          latestAnalysis.recommendedAction === 'manual_review' ||
-          latestAnalysis.recommendedAction === 'restrict'
-            ? latestAnalysis.recommendedAction
-            : 'manual_review',
-        analyzedMessageCount: latestAnalysis.analyzedMessageCount,
-      },
-    };
-  }
-
-  /**
-   * Creates an action row with admin action buttons
-   * @param userId The ID of the user the actions apply to
-   * @returns An ActionRowBuilder with buttons
-   */
-  private createActionRow(userId: string): ActionRowBuilder<ButtonBuilder> {
-    return new ActionRowBuilder<ButtonBuilder>().addComponents(
-      new ButtonBuilder()
-        .setCustomId(buildCaseAdminActionsCustomId(userId))
-        .setLabel('Admin Actions')
-        .setStyle(ButtonStyle.Primary)
-    );
-  }
-
-  private createObservedActionRows(
-    userId: string,
-    detectionEventId: string
-  ): ActionRowBuilder<ButtonBuilder>[] {
-    return [
-      new ActionRowBuilder<ButtonBuilder>().addComponents(
-        new ButtonBuilder()
-          .setCustomId(buildObservedAdminActionsCustomId(userId, detectionEventId))
-          .setLabel('Admin Actions')
-          .setStyle(ButtonStyle.Primary)
-      ),
-    ];
   }
 
   /**
@@ -1710,12 +859,15 @@ export class NotificationManager implements INotificationManager {
     const updatedEmbed = messageEmbeds.length > 0 ? EmbedBuilder.from(messageEmbeds[0]) : null;
 
     if (updatedEmbed) {
-      this.upsertVerificationActionFailureField(updatedEmbed, verificationEvent);
+      this.presentationBuilder.upsertVerificationActionFailureField(
+        updatedEmbed,
+        verificationEvent
+      );
     }
 
     await message.edit({
       allowedMentions: { parse: [] },
-      components: [this.createActionRow(verificationEvent.user_id)],
+      components: [this.presentationBuilder.createActionRow(verificationEvent.user_id)],
       ...(updatedEmbed ? { embeds: [updatedEmbed] } : {}),
     });
   }
